pyproject.toml
[tool.poetry]
name = "alembic-run_async"
version = "0.1.0"
description = ""
authors = ["Palaniyappan Ramanathan <vickypalaniappan12@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
sqlalchemy = "^2.0.31"
alembic = "^1.13.2"
python-dotenv = "^1.0.1"
asyncpg = "^0.29.0"
greenlet = "^3.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
